# Week4_ASP.NET Core 8.0 Web API/WebApi_Handson_5/Code/SwaggerDemoAPI/Models/Employee.cs
using System;
using System.Collections.Generic;

namespace SwaggerDemoAPI.Models
{
    public class Employee
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public int Salary { get; set; }
        public bool Permanent { get; set; }
        public Department Department { get; set; }
        public List<Skill> Skills { get; set; }
        public DateTime DateOfBirth { get; set; }
    }
}